+++ truffle/src/com.oracle.truffle.api.debug.test/src/com/oracle/truffle/api/debug/test/AbstractDebugTest.java
@@ -89,6 +89,10 @@ public abstract class AbstractDebugTest {
         return tester.startSession(sourceElements);
     }
 
+    protected final String getOutput() {
+        return tester.getOut();
+    }
+
     protected final Thread getEvalThread() {
         return tester.getEvalThread();
     }